Many changes related to supporting the DTM Builder. String pool and coroutine parser layers now have built-in unit-test drivers; see their main()s.
The calls from DTMBuilder to DTM still need a lot of work to
polish exactly what pools are used and what filtering goes on
during DTM build. More importantly, the whole startup sequence
needs to be properly architected -- the DTM, builder and
parser have cross-awareness of each other and someone
needs to decide who creates what and/or connects to what
in what order.

In Progress!