SAX filtering test written. Mostly works.
One unaddressed case: If the parser crashes out, either on request
(because the user issued doTerminate) or accidentally, the processing
thread is _not_ informed. This can be addressed by having the main
thread explicitly catch the exception and issue the exception through
the coroutine system as if it came from the parser... but  in the case
Scott is interested in, we would have no way to tell the event source that
it must do this!

"Quite a three-pipe problem, Watson."