Removed all dependencies of non-stree classes on stree.  RTF construction is now being done with the DTM.
Defined null Source object for DTMManager#getDTM(Source...
to mean "make me a DTM but don't construct it".

Added incremental arg to DTMManager#getDTM(Source... to tell
if it should build a CoRoutineParser and pass it to the DTM.

Added getLexicalHandler, getDTDHandler, etc., to the DTM
interface.  Ultimately needs review.

Fixed some bugs...