DMManagerDefault@getDTMHandleFromNode now tries to find an existing DOM2DTM which contains that node, rather than immediately generating a new one. The cost of searching for it is fairly high, though I've tried to avoid exhaustive search through every DOM2DTM.
Caveat: If found, there is no guarantee that this is the _best_
DTMHandle for that node. Due to the possiblility that DOM2DTM
represents only a subtree, DTMManager may have multiple DTMs
which expose that node. We could search for the one that gives
it the deepest "level"... but given how expensive each search
currently is, I don't know if we want to go there.