Implemented new isNodesetExpr and asNode methods on Expression, and getArg0AsNode and Arg0IsNodesetExpr on FunctionDef1Arg. This allows functions that just take a single node to get that node without creating an XNodeset or iterator if the expression is a simple one-step LocationPath (more complicated LocationPaths still need iterators, at least for the moment).  There will be more to come along these lines, trying to reduce the number of XObject derivatives that are are created during function evaluation.
Added sendNormalizedSAXcharacters to the FastStringBuffer, and
a param on DTM#dispatchCharactersEvents to enable normalization
of the characters.  This allows us to send normalized characters
without copying.  I think the sendNormalizedSAXcharacters may
need a little tuning.

Added tweak/hack to DTMDefaultBase which is a flag telling
if a namespace has been seen, and, if it has not, always return
immediately from getFirstNamespaceNode, which, as it turns
out, is called a lot in the course of copy operations.  We need to
come up with something better than this.  I suggested to Joe
that we use a bitset for namespaces and attributes that will
tell us if a given element has these nodes (I think it would help
for getFirstAttribute as well).