Make XSLTC on DTM completely independent of Xalan and XPath
1. Remove the reference to org.apache.xpath.objects.XMLStringFactoryImpl
in XSLTCDTMManager. A null XMLStringFactory is used when creating a SAXImpl.

2. In SAXImpl, replace the references to getStringValue() by getStringValueX() which
does not use XMLStrings.

3. In SmartTransformerFactoryImpl.java, do not explicitly reference
org.apache.xalan.processor.TransformerFactoryImpl to get rid of
the compile time dependency on Xalan.

4. Do not use reflection in XSLTCDTMManaget.newInstance(). There is no intention
to let the user specify another XSLTCDTMManager at the moment. We can revisit
this issue if there is a future need.